[option_quiet] Add support to option `--quiet` in azk commands
